refactor(react-gui): lift a row and a debounce out of the last two panes
The last two panes on the list, and both are smaller cases than the ones
before them: one component and one effect, each with a reason to be named.

DensityMapPane's `DragRow` looks like the inspector's `NumRow` and cannot
be it: the map panel writes properties by name, while `NumRow` takes a
`GenericPropEntry` and reports back through it. The plan expected these to
merge; they cannot, and the row now says so where the next reader will
look. It keeps the same realtime lifecycle either way -- preview per
frame, one commit on release, rollback on abort.

SelectionPane's live validation is a debounce with a cancelled flag, and
both halves are load-bearing: a selection compiles in C++, so asking per
keystroke would compile a dozen half-written expressions per word, and a
reply that arrives after the text moved on would mark the wrong
expression. The empty string, `*` and `none` skip the round trip entirely
-- they are always valid and are what the field holds most of the time.
It is also optimistic on failure: a failed round trip reports valid,
rather than marking a field the user has no way to fix.

Neither pane gained an exhaustive-deps warning, which is the first split
in this run that did not -- the extracted pieces take values, not setters.
